+++ eca-server/src/test/java/com/ecaservice/server/TestHelperUtils.java
@@ -124,7 +124,7 @@ public class TestHelperUtils {
     private static final String HIDDEN_LAYER = "5,8,9";
     private static final int MAX_DEPTH = 10;
     private static final int NUM_ITERATIONS = 5;
-    private static final double MIN_ERROR = 0.00001d;
+    private static final BigDecimal MIN_ERROR = BigDecimal.valueOf(0.00001d);
     private static final int NUM_IN_NEURONS = 10;
     private static final int NUM_OUT_NEURONS = 12;
     private static final double LEARNING_RATE = 0.01d;
